desktop-file: remove ICED as a category
I am working on patches to get COSMIC added as a category in [xdg-specs](https://gitlab.freedesktop.org/xdg/xdg-specs)

I've been told though, that toolkit categories (i.e. GTK, Qt, etc) are essentially deprecated, and not really useful to add.

As a result, to pass desktop file validation, we should remove ICED as a category. We can leave COSMIC, as that will (hopefully soon) be a valid category.